+++ R/multipleGroup.R
@@ -77,7 +77,7 @@
 #'   \code{\link{expand.table}}, \code{\link{key2binary}}, \code{\link{mirt.model}}, \code{\link{mirt}},
 #'   \code{\link{bfactor}}, \code{\link{multipleGroup}}, \code{\link{mixedmirt}},
 #'   \code{\link{wald}}, \code{\link{itemplot}}, \code{\link{fscores}}, 
-#   \code{\link{fitIndices}},
+#'   \code{\link{M2}},
 #'   \code{\link{extract.item}}, \code{\link{iteminfo}}, \code{\link{testinfo}}, \code{\link{probtrace}},
 #'   \code{\link{boot.mirt}}, \code{\link{imputeMissing}}, \code{\link{itemfit}}, \code{\link{mod2values}},
 #'   \code{\link{simdata}}, \code{\link{createItem}}, \code{\link{mirtCluster}}, \code{\link{DIF}}
@@ -98,6 +98,8 @@
 #' models <- mirt.model('F1 = 1-15')
 #'
 #' mod_configural <- multipleGroup(dat, models, group = group) #completely separate analyses
+#' #limited information fit statistics
+#' M2(mod_configural)
 #'
 #' mod_metric <- multipleGroup(dat, models, group = group, invariance=c('slopes')) #equal slopes
 #' #equal intercepts, free variance and means